+++ fsl_mrs/tests/test_scripts_preproc.py
@@ -33,6 +33,38 @@ def test_preproc(tmp_path):
          '--ecc', ecc,
          '--t1', t1,
          '--align_limits', '0.5', '4.0',
+         '--remove-water',
+         '--truncate-fid', '1',
+         '--overwrite',
+         '--report',
+         '--verbose'])
+
+    assert retcode == 0
+    assert (tmp_path / 'mergedReports.html').exists()
+    assert (tmp_path / 'voxel_location.png').exists()
+    assert (tmp_path / 'metab.nii.gz').exists()
+    assert (tmp_path / 'wref.nii.gz').exists()
+
+    proc_nii = mrs_io.read_FID(tmp_path / 'metab.nii.gz')
+    assert proc_nii.shape == (1, 1, 1, 4095)
+
+
+def test_depreciated_options(tmp_path):
+
+    metab = str(data / 'metab_raw.nii.gz')
+    wrefc = str(data / 'wref_raw.nii.gz')
+    wrefq = str(data / 'quant_raw.nii.gz')
+    ecc = str(data / 'ecc.nii.gz')
+
+    retcode = subprocess.check_call(
+        ['fsl_mrs_preproc',
+         '--output', str(tmp_path),
+         '--data', metab,
+         '--reference', wrefc,
+         '--quant', wrefq,
+         '--ecc', ecc,
+         '--t1', t1,
+         '--align_limits', '0.5', '4.0',
          '--hlsvd',
          '--leftshift', '1',
          '--overwrite',
@@ -47,6 +79,34 @@ def test_preproc(tmp_path):
 
     proc_nii = mrs_io.read_FID(tmp_path / 'metab.nii.gz')
     assert proc_nii.shape == (1, 1, 1, 4095)
+
+
+def test_noalign(tmp_path):
+
+    metab = str(data / 'metab_raw.nii.gz')
+    wrefc = str(data / 'wref_raw.nii.gz')
+    wrefq = str(data / 'quant_raw.nii.gz')
+    ecc = str(data / 'ecc.nii.gz')
+
+    retcode = subprocess.check_call(
+        ['fsl_mrs_preproc',
+         '--output', str(tmp_path),
+         '--data', metab,
+         '--reference', wrefc,
+         '--quant', wrefq,
+         '--ecc', ecc,
+         '--noalign',
+         '--overwrite',
+         '--report',
+         '--verbose'])
+
+    assert retcode == 0
+    assert (tmp_path / 'mergedReports.html').exists()
+    assert (tmp_path / 'metab.nii.gz').exists()
+    assert (tmp_path / 'wref.nii.gz').exists()
+
+    proc_nii = mrs_io.read_FID(tmp_path / 'metab.nii.gz')
+    assert proc_nii.shape == (1, 1, 1, 4096)
 
 
 def test_window_align_preproc(tmp_path):
@@ -66,8 +126,8 @@ def test_window_align_preproc(tmp_path):
          '--t1', t1,
          '--align_limits', '0.5', '4.0',
          '--align_window', '4',
-         '--hlsvd',
-         '--leftshift', '1',
+         '--remove-water',
+         '--truncate-fid', '1',
          '--overwrite',
          '--report',
          '--verbose'])
@@ -135,7 +195,7 @@ def test_preproc_fmrs(tmp_path):
          '--ecc', ecc,
          '--t1', t1,
          '--fmrs',
-         '--leftshift', '1',
+         '--truncate-fid', '1',
          '--overwrite',
          '--report',
          '--verbose'])
@@ -162,8 +222,8 @@ def test_already_processed(tmp_path):
              '--data', metab,
              '--reference', wref,
              '--t1', t1,
-             '--hlsvd',
-             '--leftshift', '1',
+             '--remove-water',
+             '--truncate-fid', '1',
              '--overwrite',
              '--report'],
             check=True,
@@ -218,8 +278,8 @@ def test_wref_singleton_preproc(tmp_path):
          '--ecc', tmp_path / 'ecc.nii.gz',
          '--t1', t1,
          '--align_limits', '0.5', '4.0',
-         '--hlsvd',
-         '--leftshift', '1',
+         '--remove-water',
+         '--truncate-fid', '1',
          '--overwrite',
          '--report',
          '--verbose'],